requirement.txt
flask
flask-cors
selenium
webdriver-manager
beautifulsoup4
ollama
requests
python-dotenv
cheerio